engine/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
aiosqlite>=0.20.0
pyyaml>=6.0.1
httpx>=0.27.0
anthropic>=0.39.0
python-dotenv>=1.0.1
sse-starlette>=2.1.0
websockets>=13.0
bcrypt>=4.0.0
pyjwt>=2.8.0
